+++ drlnd/common/agents/config.py
@@ -1,0 +1,11 @@
+import torch
+
+BUFFER_SIZE = int(1e5)  # replay buffer size
+BATCH_SIZE = 64         # minibatch size
+GAMMA = 0.99            # discount factor
+TAU = 1e-3              # for soft update of target parameters
+LR = 5e-4               # learning rate 
+UPDATE_EVERY = 4        # how often to update the network
+COPY_WEIGHTS_EVERY = 50 # how often to update the target network using local weights
+
+device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
